bin: move non-bin mod to subdir
