+++ src/bin/utils/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(unix)]
+pub(crate) mod profiling;
+#[macro_use]
+pub(crate) mod setup;
+pub(crate) mod signal_handler;
