+++ src/bin/utils/signal_handler.rs
@@ -15,7 +15,7 @@
 mod imp {
     use libc::c_int;
 
-    use profiling;
+    use utils::profiling;
 
     use tikv::raftstore::store::Engines;
     use tikv::util::{metrics, rocksdb_stats};
